+++ src/main/java/com/github/deeepamin/ciaid/services/resolvers/NeedsOrExtendsToJobReferenceResolver.java
@@ -23,7 +23,7 @@ public class NeedsOrExtendsToJobReferenceResolver extends SingleTargetReferenceR
 
   @Override
   public Object @NotNull [] getVariants() {
-    var isInputsString = GitlabCIYamlUtils.getInputNameFromInputsString(myElement.getText()) != null;
+    var isInputsString = GitlabCIYamlUtils.isAnInputsString(myElement.getText());
     if (isInputsString) {
       return new LookupElement[0];
     }